refactor(sms): migrate 32 SMS update arms into SmsConversationStore::update
